requirements.txt
yfinance
pandas
google-cloud-storage
numpy
matplotlib 
seaborn 
tensorflow
python-dotenv
kaleido
plotly